requirements_macos.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
altgraph==0.17.4
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
bcrypt==4.2.1
bidict==0.23.1
bitarray==3.0.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.1
# cupy-cuda12x==12.3.0  £ for CUDA 12.x, macOS does not support
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.10
Cython==3.0.12
datasets==2.21.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
docker==7.1.0
docopt==0.6.2
ecdsa==0.19.0
editdistpy==0.1.5
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
fastrlock==0.8.3
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
HyperPyYAML==1.2.2
hypothesis==6.126.0
idna==3.10
iniconfig==2.1.0
# intersystems_irispython @ file:///D:/medsync/install_iris/intersystems_irispython-5.0.1-8026-... # .whl file for Windows, cannot install in macOS
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiwer==3.1.0
joblib==1.4.2
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
# kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=... # github zip이 자주 바뀌어 설치 중단 발생
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.9.2
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.43.0
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
optuna==4.2.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
preshed==3.0.9
primePy==1.3
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyahocorasick==2.1.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
#pyarrow==15.0.2 #pyarrow==19.0.0
pyasn1==0.4.8
PyAudio==0.2.14
pycparser==2.22
pyctcdecode==0.5.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pygtrie==2.5.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.5
pytest-json==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.4.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-socketio==5.11.4
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.2
regex==2024.11.6
requests>=2.32.2 # requests==2.31.0
resampy==0.4.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semver==3.0.4
sentence-transformers==3.4.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-alignments==0.9.1
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
spacy-transformers==1.3.8
speechbrain==1.0.2
SQLAlchemy==2.0.36
sqlalchemy-iris==0.17.0
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
symspellpy==6.7.8
tabulate==0.9.0
tensorboardX==2.6.2.2
testcontainers==4.9.1
# testcontainers-iris==1.2.2  # (필요한 경우만 다시 추가)
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.2.2  # 최신 버전 중 macOS에서 지원되는 버전
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2 # torchaudio==2.6.0
torchmetrics==1.6.1
torchvision==0.17.2 # torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webrtcvad==2.0.10
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
